requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==3.4.7
Django==3.2.5
django-appconf==1.0.4
django-ckeditor==6.1.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-rest-framework==0.1.0
djangorestframework==3.12.4
jwt==1.2.0
mysqlclient @ file:///C:/np_venvs/naengpa/Scripts/mysqlclient-1.4.6-cp38-cp38-win32.whl
pilkit==2.0
Pillow==8.3.1
pycparser==2.20
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
